Fixed a bug that could lead to an infinite loop
If an expected timebound could not be found for some transition, but an
expected costbound it was marked that the time bounds changed.
Therefore this leads to an endless loop since there will always be an
unbounded general transition but we iterate nonetheless. Hence we only
marked the time bounds as changed if an expected costbound could be
found for a previous *unbound* transition.