Added an additional smt minimisation method
The new method iteratively tries to minimise the absolute values of variables.
To that end a custom binary search is used, hence it only yields optimal
results on the integer domain.
The method can for example be used in the minimisation of computed
ranking function as it works better than the minimize_absolute_old
method.
However it is not used at the moment since it doubles the runtime while
in general not leading to better asymptotic bounds.