Revert "Completely reworked program renaming"
The variables are uniquely identified by the position at which they
occur in the locations. Hence, the original renaming is *mandatory*.

To deal with ExpectedSize var goals, we assume that var refers to a
global and not *local* variable such that its position is defined by the
position of var in the (VARIABLES ...) string.

This reverts commit 5f87f32aa9c263f07bc56de47e5f13fb993c58d7.