Smarter computation of Ranking Functions & Performance Improvement
We now only compute Ranking Functions that can be used to compute a
finite time bound.
To that end, the non-increasing set is chosen such that the time bound
of all entry transitions is finite, and such that all variables appearing
with a non-zero coefficient have bounded size (at the given location).

To incorporate newly found bounds we have to recompute all ranking
functions after one pass of the analysis.

Furthermore, this commit contains performance improvements by
avoiding the (costly) recomputation of entry transitions in many cases.