Improve performance of MPRF
Instead of backtracking over the complete SCC to find the maximal
set of non-increasing transitions, we now compute pairings
(tset1, tset2) of two transition sets.

Here, tset1 is a subset of the transitions we wish to make decreasing.
tset2 is (for each transition in tset1) the "closure" such that
all pre-transitions of this set are time bounded.

As a result, in many cases the size of the possible non-increasing
transitions we still have to find using backtracking is drastically
decreased.
(Note that in the worst case we loop over the powerset so the time
complexity of the backtracking is exponential in the number of possible non-increasing
transitions)