Accomodate for non-local effects of apply_cfr
Analysis.apply_cfr with preprocessing might have unintented side effects outside of the scope of the current SCC.

Consider example `TPDB/Complexity_ITS/Brockschmidt_16/T2/apchild-accepted-fail.koat`.
After CFR one might be able to show insatisfiability of transition f105 → f120.
As a result this transition (and additionally the subsequent SCC) might be cut.
Hence, when this SCC is scheduled for analysis it is not contained in the analysed program anymore.