Analysis: Compute RVG per SCC
Note that here we have to consider trivial transitions (SCC-connecting
transitions) as well.
Since `Program.sccs` views SCCs as set of transitions they do not occur here.
Hence, in the analysis we view SCCs as set of locations instead.

For a given SCC we then compute the RVG as the RVG that considers all
the transitions inside the SCC as well as the SCC-connecting transitions entering or leaving this SCC.

For Trivial Size Bounds it is then important to compute the
`pre_transitions` from the program instead of using `RVG.pre` as
entering transitions do not have a pre in the rvg and are hence assigned
a size bound of 0 which is unsound in general.