Preprocess again after CFR (Read Notes)
Preprocessing once more *after* CFR might lead to the cut of program locations.
Since the SCCs to be analysed are computed from the original program we can't be sure that these still exist in the preprocessed refined program, as the effects of preprocessing are currently not limited to the refined CFR.

Hence, in `improve_scc` we have to consider the intersection of the SCC locations and all locations present in the current program iff CFR is enabled.