NativePartialEvaluation: Fix bug
Correctly deal with unfolding leading to a trivial version <ℓ,⊤>.

We use the Hashtbl from version to location names to also track already seen versions.
Before we'd all trivial versions to this Hashtbl too ensure that trivial versions map to original locations.
This however, meant that encountering a trivial version lead to a halt of the partial evaluation.

E.g. consider a program `ℓ₀() → ℓ₁(), ℓ₁() → ℓ₂(), ℓ₂() → ℓ₁()` where we want to refine the complete loop.
Then, one would unfold <ℓ₁,⊤> to <ℓ₂,⊤> and stop evaluation.
However, the final transition ℓ₂ → ℓ₁ is not an exit transition and is hence not readded to the program leading to removal of the loop.

Now, we only add versions to the hash table that have been encountered in the partial evaluation.
To make this work we generate the "new location name" ℓ for trivial versions <ℓ,⊤>.